Introduce Updateable interface for converting datetime to HTTP string
The idea is to have the string representation in a single place so that
we can add response header based on being instance of that interface.
